refactor(web): server/auth の寄せ集めヘルパー置き場を解消
route を持たず registrations / sessions から共有されるだけの薄い
server/auth/ をやめ、中身を適切な場所へ移す。

- captcha.ts / captcha.test.ts を server 直下フラットへ移動
- authInputValidator は各 handler/create.ts へ co-locate（validator を
  ハンドラで定義・export する既存慣習に揃え、validators.ts を廃止）
- 空になった server/auth/ を削除
