+++ application/apps/web/src/server/registrations/route.ts
@@ -1,8 +1,7 @@
 import { Hono } from 'hono'
 import type { Env } from '@/env'
 import rateLimiter from '@/middleware/rate-limiter'
-import { authInputValidator } from '@/server/auth/validators'
-import create from './handler/create'
+import create, { authInputValidator } from './handler/create'
 
 const app = new Hono<Env>().post('/', rateLimiter, authInputValidator, create)
