+++ application/apps/web/src/server/sessions/route.ts
@@ -2,8 +2,7 @@ import { Hono } from 'hono'
 import type { Env } from '@/env'
 import { authenticator } from '@/middleware/authenticator'
 import rateLimiter from '@/middleware/rate-limiter'
-import { authInputValidator } from '@/server/auth/validators'
-import create from './handler/create'
+import create, { authInputValidator } from './handler/create'
 import getCurrentSession from './handler/current'
 import destroy from './handler/destroy'
